Refactor commands. Thank you bro
